refactor: move agendamento module classes
