+++ src/main/java/dao/ConnectionFactory.java
@@ -1,4 +1,4 @@
-package agendamento.dao;
+package dao;
 
 import java.io.IOException;
 import java.io.InputStream;
