+++ src/main/java/model/JobStep.java
@@ -1,14 +1,14 @@
-package agendamento.model;
-
-import java.time.Instant;
+package model;
 
 /**
- * Representation of a job loaded from the database.
+ * Step definition belonging to a job.
  */
-public class Job {
+public class JobStep {
     private long id;
+    private long jobId;
+    private int ordem;
     private String nome;
-    private boolean ativo;
+    private boolean habilitado;
     private String handlerClass;
     private String handlerMethod;
     private boolean handlerStatic;
@@ -16,19 +16,24 @@ public class Job {
     private String workingDir;
     private String jvmArgs;
     private String envVars;
-    private ConcurrencyPolicy policy = ConcurrencyPolicy.ALLOW;
+    private String parametros;
+    private String condicaoExpr;
+    private boolean continueOnFail;
+    private int timeout;
     private int maxRetries;
     private int retryBackoff;
-    private int timeout;
-    private Instant createdAt;
 
-    // getters and setters
+    // getters and setters omitted for brevity
     public long getId() { return id; }
     public void setId(long id) { this.id = id; }
+    public long getJobId() { return jobId; }
+    public void setJobId(long jobId) { this.jobId = jobId; }
+    public int getOrdem() { return ordem; }
+    public void setOrdem(int ordem) { this.ordem = ordem; }
     public String getNome() { return nome; }
     public void setNome(String nome) { this.nome = nome; }
-    public boolean isAtivo() { return ativo; }
-    public void setAtivo(boolean ativo) { this.ativo = ativo; }
+    public boolean isHabilitado() { return habilitado; }
+    public void setHabilitado(boolean habilitado) { this.habilitado = habilitado; }
     public String getHandlerClass() { return handlerClass; }
     public void setHandlerClass(String handlerClass) { this.handlerClass = handlerClass; }
     public String getHandlerMethod() { return handlerMethod; }
@@ -43,14 +48,16 @@ public class Job {
     public void setJvmArgs(String jvmArgs) { this.jvmArgs = jvmArgs; }
     public String getEnvVars() { return envVars; }
     public void setEnvVars(String envVars) { this.envVars = envVars; }
-    public ConcurrencyPolicy getPolicy() { return policy; }
-    public void setPolicy(ConcurrencyPolicy policy) { this.policy = policy; }
+    public String getParametros() { return parametros; }
+    public void setParametros(String parametros) { this.parametros = parametros; }
+    public String getCondicaoExpr() { return condicaoExpr; }
+    public void setCondicaoExpr(String condicaoExpr) { this.condicaoExpr = condicaoExpr; }
+    public boolean isContinueOnFail() { return continueOnFail; }
+    public void setContinueOnFail(boolean continueOnFail) { this.continueOnFail = continueOnFail; }
+    public int getTimeout() { return timeout; }
+    public void setTimeout(int timeout) { this.timeout = timeout; }
     public int getMaxRetries() { return maxRetries; }
     public void setMaxRetries(int maxRetries) { this.maxRetries = maxRetries; }
     public int getRetryBackoff() { return retryBackoff; }
     public void setRetryBackoff(int retryBackoff) { this.retryBackoff = retryBackoff; }
-    public int getTimeout() { return timeout; }
-    public void setTimeout(int timeout) { this.timeout = timeout; }
-    public Instant getCreatedAt() { return createdAt; }
-    public void setCreatedAt(Instant createdAt) { this.createdAt = createdAt; }
 }
