+++ src/main/java/model/StepRun.java
@@ -1,4 +1,4 @@
-package agendamento.model;
+package model;
 
 import java.time.Instant;
